requirements.txt
aiogram==2.18
bs4
pycairo
python-dotenv